requirements.txt
selenium
webdriver_manager
feedparser
notion-client
python-dotenv
google-generativeai
